feat(storage): TODO#11 — filesystem command surface (cmd_files)
Add cmd_files.cpp with a full 'files' command surface:
  - files / files status  — both backends' mount state and storage info
  - files use <flash|sd>  — switch active backend (session-level, defaults to flash)
  - files ls [path]       — directory listing with type and size (default /)
  - files cat <path>      — print file contents, capped at 8 KB with truncation note
  - files info <path>     — type and size for file or directory
  - files rm / mkdir / rmdir <path>
  - files write / append <path> <text>
  - files format confirm  — format active backend (requires explicit confirmation)

Fix SdmmcGetStorageInfo: change size_t* to uint64_t* to avoid 32-bit overflow
on cards larger than 4 GB; replace statvfs (unavailable on RISC-V newlib) with
esp_vfs_fat_info(); use %lu with pre-divided MB value instead of %llu (newlib
does not support %llu on ESP32-P4).

Wire CommandRegisterFiles() into CommandRegisterBuiltins().

+++ src/system/adapters/sdmmc_adapter.h
@@ -48,7 +48,7 @@ bool SdmmcRmdir(const char* path);
 
 // Returns basic metadata. Pass nullptr for out parameters you do not need.
 bool SdmmcStat(const char* path, bool* out_is_dir, size_t* out_size);
-bool SdmmcGetStorageInfo(size_t* out_total, size_t* out_used);
+bool SdmmcGetStorageInfo(uint64_t* out_total, uint64_t* out_used);
 bool SdmmcList(const char* path, SdmmcListCallback callback, void* user);
 
 bool SdmmcReadBytes(const char* path, uint8_t* out, size_t out_size, size_t* out_len);
